# routes/feedback/actions.js
/**
 * @exports an express router
 * ACTIONS are the model concerned by these routes
 * This is part of the feedback module
 * actions are generated by customers to indicate their intent
 * examples of possible action types:
 * 'WILLING_TO_REVIEW'
 * 'GOOGLE_REVIEW'
 * 'SENT_NEGATIVE_FEEDBACK'
 * 'INTERESTED_IN_BEING_CONTACTED'
 */
var router = require("express").Router(),
  mongoose = require('mongoose'),
  { CrudFactory, SendFunctions } = require('utilities_middleware'),
  Actions = CrudFactory(mongoose.model('Action')),
  { sendData } = SendFunctions,
  { permitRoles } = require("auth_module/middleware"),
  { adminAllClientSelf } = require(process.cwd() + "/utils/sequences");

/**
 * @function gets an action create by a user
 * Role restrictions:
 * - client can view actions taken by their customers
 * - admin can view actions from all customers
 * @param query.user      the user id of a business
 */
router.get(`/action`,
  adminAllClientSelf("query.auth"),
  Actions.read(),
  sendData);

/**
 * @function gets a list of actions created by a user
 * Role restrictions:
 * - client can view actions taken by their customers
 * - admin can view actions from all customers
 * @param query.user      the user id of a business
 */
router.get(`/action/list`,
  adminAllClientSelf("query.auth"),
  Actions.list,
  sendData
);

/**
 * @function any customer can create an action, that's why this route isn't restricted
 * @param req.body.type     a string literal
 * @param req.body.user     the business for which it's relevant
 * @param req.body.customer the id of the customer
 */
router.post(`/action`,
  Actions.create(),
  sendData);

/**
 * @function after they happen, clients and customers shouldn't change actions
 * Role restrictions:
 * - admin only
 */
router.put(`/action`,
  permitRoles("admin"),
  Actions.update,
  sendData);

/**
 * @function deletes actions, clients and customers can't do this
 * Role restrictions:
 * - admin only
 */
router.delete(`/action`,
  permitRoles("admin"),
  Actions.remove(),
  sendData);

module.exports = function(app) {
  app.use(router);
};