# swagger.yaml
swagger: '2.0'
info:
  description: How to use the Boastable server.
  version: 1.0.0
  title: Boastable API
  contact:
    email: ryan@fueledonbacon.com
host: localhost:3000
basePath: /
...
schemes:
  - http
paths:
  /login:
    post:
      tags:
        - auth
      summary: Create a new business/user profile
      description: ''
      operationId: login
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Allows a valid user to login
          required: true
          schema:
            $ref: '#/definitions/CreateAuthBody'
      responses:
        '200':
          description: JWT token
          schema:
            $ref: '#/definitions/TokenResponse'
        '401':
          description: Invalid credentials
...
  /auth/list:
    get:
      tags:
        - auth
      summary: Get a list of users
      description: 'this is an admin only function. lists all Auth documents'
      operationId: listAuthDocuments
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          description: an authorization header
          example: bearer [insert JWT token here]
          required: true
          type: string
        - name: query
          in: query
          description: 'any key another user document to filter for'
          required: false
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Membership'
        '400':
          description: Invalid username supplied
        '404':
          description: User not found
  /auth:
    post:
      tags:
        - auth
      summary: Create auth document
      description: Clients create their own customers
      operationId: createAuth
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          description: an authorization header
          example: bearer [insert JWT token here]
          required: true
          type: string
        - in: body
          name: body
          description: Created auth document
          required: true
          schema:
            $ref: '#/definitions/Auth'
      responses:
        '200':
          description: OK
    get:
      tags:
        - auth
      summary: Get customer
      description: ''
      operationId: getAuth
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          description: an authorization header
          example: bearer [insert JWT token here]
          required: true
          type: string
        - name: id
          in: query
          description: get auth document for a client
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Auth'
        '400':
          description: Invalid username supplied
        '404':
          description: Auth not found
    put:
      tags:
        - auth
      summary: Update customer data
      description: This can only be done by the logged in user.
      operationId: updateAuth
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          description: an authorization header
          example: bearer [insert JWT token here]
          required: true
          type: string
        - name: id
          in: query
          description: ID of the customer that needs to be updated
          required: true
          type: string
        - in: body
          name: body
          description: Updated customer object
          required: true
          schema:
            $ref: '#/definitions/Auth'
      responses:
        '200':
          description: OK
        '400':
          description: Invalid user supplied
        '404':
          description: Auth not found
    delete:
      tags:
        - auth
      summary: Delete auth document
      description: Only an admin can do this
      operationId: deleteAuth
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          description: an authorization header
          example: bearer [insert JWT token here]
          required: true
          type: string
        - name: id
          in: query
          description: A customer that needs to be deleted
          required: true
          type: string
      responses:
        '200':
          description: OK
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '404':
          description: Auth not found
...
  /sms:
    post:
      tags:
        - communication
      summary: Send an SMS
      description: Client and Admin role users can use this route to send SMS messages
      operationId: createSMS
      produces:
        - application/xml
        - application/json
      parameters:
        - in: body
          name: body
          description: An E.164 phone number and message
          required: true
          schema:
            $ref: '#/definitions/SMS'
      responses:
        default:
          description: successful operation
  /reset-password:
    post:
      tags:
        - communication
      summary: Sends an email to the user with a link to reset their password
      description: Any user can use this endpoint.
      operationId: resetPassword
      produces:
        - application/xml
        - application/json
      parameters:
        - in: body
          name: body
          description: An email address
          required: true
          schema:
            $ref: '#/definitions/ResetPasswordRequest'
      responses:
        default:
          description: OK
  /comment:
    post:
      tags:
        - communication
      summary: Sends an email to boastable admin containing a comment from a user
      description: Any client can use this endpoint.
      operationId: sendComment
      produces:
        - application/xml
        - application/json
      parameters:
        - name: Authorization
          in: header
          description: an authorization header
          example: bearer [insert JWT token here]
          required: true
          type: string
        - in: body
          name: body
          description: A comment from a client
          required: true
          schema:
            $ref: '#/definitions/CommentRequestBody'
      responses:
        default:
          description: OK
...
  /action/list:
    get:
      tags:
        - actions
      summary: List actions
      description: 'Returns all actions from customers a user is allowed to see or query. Client results will be filtered to only return actions from customers they created'
      operationId: listActions
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          description: an authorization header
          example: bearer [insert JWT token here]
          required: true
          type: string
        - name: query
          in: query
          description: Any key from the customer document
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Action'
        '400':
          description: Invalid username supplied
        '404':
          description: Action not found
  /action:
    post:
      tags:
        - actions
      summary: Create action
      description: This can only be done by the logged in user.
      operationId: createAction
      produces:
        - application/xml
        - application/json
      parameters:
        - in: body
          name: body
          description: Created user object
          required: true
          schema:
            $ref: '#/definitions/Action'
      responses:
        default:
          description: successful operation
    get:
      tags:
        - actions
      summary: Get actions
      description: ''
      operationId: getAction
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          description: an authorization header
          example: bearer [insert JWT token here]
          required: true
          type: string
        - name: user
          in: query
          description: ObjectId string for the desired user (business ID)
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Action'
        '400':
          description: Invalid username supplied
        '404':
          description: User not found
    put:
      tags:
        - actions
      summary: Updated action
      description: Only an admin can do this.
      operationId: updateAction
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          description: an authorization header
          example: bearer [insert JWT token here]
          required: true
          type: string
        - name: id
          in: query
          description: action that need to be updated
          required: true
          type: string
        - in: body
          name: body
          description: Updated action object
          required: true
          schema:
            $ref: '#/definitions/Action'
      responses:
        '400':
          description: Invalid user supplied
        '404':
          description: User not found
    delete:
      tags:
        - actions
      summary: Delete actions
      description: Only an admin can do this
      operationId: deleteAction
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          description: an authorization header
          example: bearer [insert JWT token here]
          required: true
          type: string
        - name: id
          in: query
          description: An action that needs to be deleted
          required: true
          type: string
      responses:
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '404':
          description: Action not found
  /answer/list:
    get:
      tags:
        - answers
      summary: List answers
      description: 'Returns all answers from customers a user is allowed to see or query. Client results will be filtered to only return Customers they created'
      operationId: listAnswers
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          description: an authorization header
          example: bearer [insert JWT token here]
          required: true
          type: string
        - name: query
          in: query
          description: Any key from the customer document
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Answer'
        '400':
          description: Invalid username supplied
        '404':
          description: Action not found
  /answer:
    post:
      tags:
        - answers
      summary: Create answer
      description: This can only be done by the logged in user.
      operationId: createAnswer
      produces:
        - application/xml
        - application/json
      parameters:
        - in: body
          name: body
          description: Created user object
          required: true
          schema:
            $ref: '#/definitions/Answer'
      responses:
        default:
          description: successful operation
    get:
      tags:
        - answers
      summary: Get answer
      description: 'Gets the answer given by a customer'
      operationId: getAnswer
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          description: an authorization header
          example: bearer [insert JWT token here]
          required: true
          type: string
        - name: customer
          in: query
          description: The customer ID of the user who gave the answer
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Answer'
        '400':
          description: Invalid username supplied
        '404':
          description: User not found
    put:
      tags:
        - answers
      summary: Update answer
      description: This can only be done by the logged in user.
      operationId: updateAnswer
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          description: an authorization header
          example: bearer [insert JWT token here]
          required: true
          type: string
        - name: id
          in: query
          description: ID of the answer to be updated
          required: true
          type: string
        - in: body
          name: body
          description: Updated answer
          required: true
          schema:
            $ref: '#/definitions/Answer'
      responses:
        '200':
          description: OK
        '400':
          description: Invalid user supplied
        '404':
          description: Answer not found
    delete:
      tags:
        - answers
      summary: Delete answer
      description: Only an admin can do this
      operationId: deleteAnswer
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          description: an authorization header
          example: bearer [insert JWT token here]
          required: true
          type: string
        - name: id
          in: query
          description: Delete an answer given by a customer
          required: true
          type: string
      responses:
        '200':
          description: OK
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '404':
          description: Action not found
...
definitions:
  Action:
    type: object
    required:
      - customer
      - user
      - action
    properties:
      _id:
        type: string
        description: ObjectId
        example: 5a2cfe672ebedf740dac80c8
      customer:
        type: string
        description: ObjectId
        example: 5a2cfe672ebedf740dac80c8
      auth:
        type: string
        description: ObjectId, the client authorized to see this customer's information
        example: 5a2cfe672ebedf740dac80c8
      action:
        type: string
        description: string literal indicating the customer action taken
        example: WILLING_TO_REVIEW
  Auth:
    type: object
    required:
      - email
      - password
    properties:
      _id:
        type: string
        description: ObjectId
        example: 5a2cfe672ebedf740dac80c8
      email:
        type: string
        example: example@user.com
      password:
        type: string
        description: hash of plaintext password
  CreateAuthBody:
    type: object
    required:
      - email
      - password
    properties:
      email:
        type: string
        description: A valid email address
        example: example@user.com
      password:
        type: string
        description: hash of plaintext password
        example: pass123asdf
  TokenResponse:
    type: object
    properties:
      token:
        type: string
        example: 'eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiI1YjZhMDY2NzNlNmQwZDU1NGNlNDZjZWMiLCJlbWFpbCI6Im5ld0BzdWJzY3JpYmVyLmNvbSIsInJvbGUiOiJjbGllbnQifQ.MHnMzbSPt9Xpyv3kiAeEqTBksj8OXUX_l_Gb9lDZFVc'
...
  Membership:
    type: object
    required:
      - customerId
    properties:
      _id:
        type: string
        description: Mongo document ObjectId
        example: 5a2cfe672ebedf740dac80c8
      auth:
        type: string
        description: Mongo document ObjectId
        example: 5a2cfe672ebedf740dac80c8
      customerId:
        type: string
        description: Stripe customer ID
        example: cus_DJFcpaSVVzBX8c
...
  Answer:
    type: object
    required:
      - customer
      - user
      - answer
      - question
    properties:
      _id:
        type: string
        description: ObjectId
        example: 5a2cfe672ebedf740dac80c8
      customer:
        type: string
        description: ObjectId
        example: 5a2cfe672ebedf740dac80c8
      auth:
        type: string
        description: ObjectId, the client authorized to see this customer's information
        example: 5a2cfe672ebedf740dac80c8
      question:
        type: string
        description: ObjectId
        example: 5a2cfe672ebedf740dac80c8
      answer:
        type: boolean
        description: A true false answer
        example: true
...
  ResetPasswordRequest:
    type: object
    required:
      - email
    properties:
      email:
        type: string
        description: An email address matching an Auth document
        example: business@boastable.app
  CommentRequestBody:
    type: object
    required:
      - comment
    properties:
      comment:
        type: string
        description: A comment from the help section of the website
        example: How do I create question?
...
  SMS:
    type: object
    required:
      - to
      - message
    properties:
      to:
        type: string
        description: An E.164 formatted phone number
        example: +15615236666
      message:
        type: string
        description: ObjectId reference to an Auth document
        example: 5a2cfe672ebedf740dac80c8